sugar_rush_agent/requirements.txt
streamlit
google-generativeai
google-adk
scikit-learn==1.7.2
pandas
numpy
